+++ pastell-core/type-dossier/TypeDossierPersonnaliseDirectoryManager.class.php
@@ -29,15 +29,28 @@ class TypeDossierPersonnaliseDirectoryManager
      * @param TypeDossierProperties $typeDossierData
      * @throws Exception
      */
-    public function save($id_t, TypeDossierProperties $typeDossierData)
+    public function save(int $id_t, TypeDossierProperties $typeDossierData): void
     {
         $type_dossier_directory = $this->getTypeDossierPath($id_t);
+        $this->saveToDir($type_dossier_directory, $typeDossierData);
+    }
+
+    /**
+     * @param $type_dossier_directory
+     * @param TypeDossierProperties $typeDossierData
+     * @throws Exception
+     */
+    public function saveToDir($type_dossier_directory, TypeDossierProperties $typeDossierData, string $input_file_path = ''): void
+    {
         $filesystem = new Filesystem();
         if (! $filesystem->exists($type_dossier_directory)) {
             $filesystem->mkdir($type_dossier_directory);
         }
 
         $type_dossier_definition_content = $this->typeDossierTranslator->getDefinition($typeDossierData);
+        if ($input_file_path) {
+            $type_dossier_definition_content['studio_definition'] = base64_encode(file_get_contents($input_file_path));
+        }
 
         $this->ymlLoader->saveArray(
             $type_dossier_directory . "/" . FluxDefinitionFiles::DEFINITION_FILENAME,
@@ -50,7 +63,7 @@ class TypeDossierPersonnaliseDirectoryManager
      * @return string
      * @throws TypeDossierException
      */
-    public function getTypeDossierPath($id_t)
+    public function getTypeDossierPath($id_t): string
     {
         $info = $this->typeDossierSQL->getInfo($id_t);
         if (! $info) {
@@ -95,7 +108,7 @@ class TypeDossierPersonnaliseDirectoryManager
      * @param $id_type_dossier_source
      * @return string
      */
-    private function getPathToTypeDossier($id_type_dossier_source)
+    private function getPathToTypeDossier($id_type_dossier_source): string
     {
         return $this->workspace_path . "/" . self::SUB_DIRECTORY . "/module/$id_type_dossier_source";
     }