+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Il est possible de déposer en GED en utilisant l'identifiant du document Pastell comme nom de répertoire #1388
 - Il est possible de configurer le connecteur Fast parapheur pour ne pas supprimer le dossier sur le parapheur après un refus #1390
 - Simplification de l'utilisation du docker pes-viewer et mise à jour de celui-ci #1371 #1373
+- ajout de la commande `app:studio:make-module` 
 
 ## Correction